arrumei os leaks

+++ lab04/exerc_02.c
@@ -23,7 +23,7 @@ int main(){
         char c;
         while ((c = getchar()) != '\n') {}
 
-        printf("entrada %d, conteudo %c, endereco: %u \n", i, vetor[i], (unsigned int)&vetor[i]);
+        printf("entrada %d, conteudo %c, endereco: %lld \n", i, vetor[i], (long long int)&vetor[i]);
 
     }
